Server stores service JARs and MSDLs in local database
The server now uses a database instead of the file system to store
uploaded service JARs. The content of the JARs as well as the MSDL is
dynamically loaded in-memory which also helps reducing IO. This is also
a first step to a consistent abstract persistence layer which can be
used by all services. Additionally this can be used to create a fully
persistent server experience which takes care of reinstalling/restarting
services after a server restart.
after